[runtime] Enforce storage error and size boundaries

+++ runtime/src/utils/buffer/paged/view.rs
@@ -141,7 +141,7 @@ impl<B: Blob> View<'_, B> {
             bufs.truncate(0);
             return Ok((bufs, 0));
         }
-        let available = (self.size.saturating_sub(offset) as usize).min(len);
+        let available = self.size.saturating_sub(offset).min(len as u64) as usize;
         if available == 0 {
             return Err(Error::BlobInsufficientLength);
         }
